Перестроена структура файлов
